Rudimentary API for programmatic targeting
Many commands implicitly take a target as an argument, and collect that
argument only via the interactive direction chooser. Because of this,
the only way to script these commands (e.g. in autofight.lua) is to send
commands to the interactive direction chooser interface. As the
direction chooser has become more widgetified, this has become heavier
and heavier to do, leading to noticeable lag in autofire. At the same
time, interacting with this UI programmatically has long been a source
of hard-to-solve bugs and brittleness. For these reasons, this commit
adds a first pass implementation of supplying a target directly to such
commands, and uses it for autofire and autoreach. On the c++ side, there
is now a non-interactive way of calling the direction chooser that still
uses all the error checking and messaging. On the lua side, there is a
way of supplying (so far, just a few) commands directly with a
player-relative coordinate. This commit leaves automagic untouched (and
regular melee combat doesn't use the direction chooser).